+++ src/ReaderPlugin.php
@@ -89,6 +89,8 @@ class ReaderPlugin implements ProfilePluginInterface
 
             $profiles[$this->category][$name] = $this->consolidate($profile);
         }
+        // create profiles for alias
+        // note that if a profile have the same name of the alias, it will be overwritten
         foreach ($this->aliases as $alias => $profileName) {
             if (isset($profiles[$this->category][$profileName])) {
                 $profiles[$this->category][$alias] = $profiles[$this->category][$profileName];